feat(euro-process): add enum for determining which store can serve the browser
